Move common types to seperate file
