+++ common/types.ts
@@ -1,0 +1,15 @@
+type Company = {
+  id: number;
+  name: string;
+  address: string;
+  krs: string;
+};
+
+type Report = {
+  id: number;
+  dateFrom: string;
+  dateTo: string;
+  status: string;
+};
+
+export type { Company, Report };
